Renaming CFTool -> Perun. This closes #3.
