+++ .gitignore
@@ -1,6 +1,3 @@
-# Generated source code:
-cfspecification/specification/CloudFormationResourceSpecification.go
-
 # IDE or editor specific files:
 .idea
