+++ Makefile
@@ -5,14 +5,14 @@ TEST?=./...
 all: get-deps code-analysis test
 
 config-install:
-	mkdir -p "$(HOME)/.config/cftool"
-	cp defaults/main.yaml "$(HOME)/.config/cftool/main.yaml"
+	mkdir -p "$(HOME)/.config/perun"
+	cp defaults/main.yaml "$(HOME)/.config/perun/main.yaml"
 
 get-deps:
 	go get -t -v .
 	go get github.com/stretchr/testify
 	go install github.com/stretchr/testify
-	go install github.com/Appliscale/cftool
+	go install github.com/Appliscale/perun
 
 code-analysis: get-deps
 	go tool vet -v .
